Modify frame-dealer-random to support emacs daemon
When emacs is executed in daemon mode and make first frame on any
display, (display-pixel-width) return invalid value. It's because no
display assigned at the moment frame is to be created.
Therefore (display-pixel-width) is changed to be called with an
argument indicating display which is assigned to frame.

Currently, this change take effect when frame-dealer-use-hook is non-nil.