Support dealing first frame on server-mode
When using advice for setting frame position, frame-dealer make
model-frame and delete it before original make-frame call.
Therefore, when emacs is executed as server-mode and the target frame
is required by emacsclient command (i.e. The target frame is the first
frame on the target display), deleting model-frame causes killing
client and the target frame is remained as orphan.

Now, to avoid this behavior, frame-dealer abort deleting model-frame when
model-frame is the single frame linked with any client process.
Instead, frame-dealer register model-frame to
frame-dealer--model-frames, and delete it when original make-frame is
completed.